Stop workshop layouts getting squished
On medium viewports the content column ends up way
too small because of the empty space left for the
(nonexistent) sidebar nav.

Workshops don't have a sidebar, so we can just not
use this layout for them.

This is a quick fix so the content is no longer
centre-aligned, but I think that's fine as it's
way more usable for people with medium screens.